Create Hangman_v2
# Word Guessing Game in C

Welcome to the Word Guessing Game in C! This simple console-based game is a hangman-style challenge where players guess a word based on its length and the presence of specific letters.

## Features

- **Dictionary Build:** Users can input a number of words to add to the game's dictionary file (dictionary.txt). The program dynamically allocates memory for the entered words..
- **Interactive Gameplay:** Users specify the length of the word they want to guess. The program filters words from the dictionary based on the specified length. Users guess individual letters, and the program dynamically sorts letters based on their frequency in the filtered words. The game continues until the user either correctly identifies the word or reaches the maximum allowed incorrect guesses.

## Contributing

Contributions are encouraged! If you have suggestions, bug reports, or would like to add new features, feel free to open an issue or submit a pull request.